upload of initial Sound Project
Uploaded the sound files along with a CSV file that will allow for easy loading into our model. also added the python files for an easy model. after the model trains it will then have some checks to see if the model is efficent. also implemented a mic listening system to take input from the user. in SoundClass-Working.ipynb